Add a simpler mechanism for tracing non-deterministic functions
Summary:
Original Author: neildhar@fb.com
Original Git: c1c6a18f398dabdabe252202baff34ef7d1b0fd6

Add a new mechanism for tracing non-deterministic code that is primarily
in JavaScript. It uses a single HostFunction that acts as a trampoline,
taking the non-deterministic JS function as its argument, and calling it
without tracing.

This is much simpler than the current mechanism, and avoids the kind of
bugs that were identified in the previous diff.

However it does have the disadvantage of mixing references coming from
the tracing runtime with references coming from the base runtime. This
is safe, since it's in the implementation of TracingRuntime, where we
know the references are interchangeable, but it's still a drawback.

Convert `Math.random` to this new mechanism as a proof of concept.
Subsequent diffs will convert the rest.

Original Reviewed By: fbmal7

Original Revision: D38969993

Reviewed By: neildhar

Differential Revision: D41240847

fbshipit-source-id: 61cd5b0dc94ab0b4ea8bb2c104a8ed4fc372bea8